Implement merge and reduce in join operation
The MERGE and REDUCE operations make it easy to
evaluate the result of applying the join operation
in a sequence of relations.

The MERGE operation consists in fusioning two sequences
of integers to form a single sequence.

The REDUCE operations consists in taking the result of a
MERGE operation and eliminate duplicates.

The JOIN operation now returns a new relation whose tuples have length
equals the length of the vector that results from merging and reducing
the two vectors encoding variables.